static: add format string and its args checking
tt_sprintf, tt_snprintf, and tt_vsnprintf now have compile-time
checks for their arguments fitting the format string.

+++ src/box/box.cc
@@ -1115,7 +1115,7 @@ box_check_memory_quota(const char *quota_name)
 		return size;
 	diag_set(ClientError, ER_CFG, quota_name,
 		 tt_sprintf("must be >= 0 and <= %zu, but it is %lld",
-		 QUOTA_MAX, size));
+			    QUOTA_MAX, (long long)size));
 	return -1;
 }
